Added /etc/init.d/openwifi for automatic interface bring-up by loading in kernel modules and /etc/board.d/openwifi to allow lan to use dhcp client
Using the AutoLoad method worked, but it provides no clear/straightforward path to call the ./rf_init_11n.sh file (see wgd.sh in openwifi). Calling this file is necessary to enable the fir filters, otherwise no AP will become visible.

A straightforward solution is to create a openwifi service under /etc/init.d, which calls the wgd.sh script under /openwifi/user_space, the openwifi sources are installed in the root folder anyway.

For OpenWiFi, it makes sense to use dhcp client instead of a static ip on the eth port.
This is done by a file in /etc/board.d/01_openwifi